requirements.txt
requests
Flask
flask_cors
PyYAML
numpy
Pillow
opencv-python-headless
imutils
matplotlib
gdown
vietocr
seaborn
gunicorn
loguru